style: 공용컴포넌트 폴더 생성
fix: 한 개의 파일에 여러개 컴포넌트로 인한 오류 수정
