+++ src/api/supabase/index.ts
@@ -1,5 +1,7 @@
 export * from "./auth";
+export * from "./commentData";
 export * from "./interiorBookmark";
 export * from "./mypageData";
 export * from "./postBookmark";
+export * from "./postData";
 export * from "./supabaseClient";
