+++ src/api/supabase/postData.ts
@@ -25,7 +25,7 @@ export const fetchPostData = async () => {
 
 // post
 export const createPostHandler = async (postData: Tables<"POSTS", "Insert">) => {
-  const {data,error} = await supabase.from("POSTS").insert(postData).select();
+  const { data, error } = await supabase.from("POSTS").insert(postData).select();
   console.log('error :', error);
   console.log('data :', data);
 };
