+++ src/components/common/button/Button.tsx
@@ -1,8 +1,8 @@
 import React, { type ButtonHTMLAttributes } from "react";
 
-interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {}
+interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> { }
 
-const Button: React.FC<ButtonProps> = ({ children, ...props }) => {
+export const Button: React.FC<ButtonProps> = ({ children, ...props }) => {
   return (
     <button
       className="w-auto px-8 py-2 m-1 text-sm text-white transition duration-300 bg-black shadow-md rounded-3xl hover:bg-opacity-70"
@@ -12,5 +12,3 @@ const Button: React.FC<ButtonProps> = ({ children, ...props }) => {
     </button>
   );
 };
-
-export default Button;
