+++ src/components/common/input/Input.tsx
@@ -1,0 +1,1 @@
+export const Input = () => { };
