+++ src/components/common/searchBar/SearchBar.tsx
@@ -1,36 +1,3 @@
-type dataType = "post" | "comment" | "bookmark" | "like"
-
-interface Props {
-  type: dataType
-}
-
-export const MypageTitle = () => {
-  return (
-    <div className="w-full  text-center pb-[24px]">
-      <h3 className="text-[32px] font-normal leading-[130%]">마이페이지</h3>
-    </div>
-  )
-}
-export const MypageSubTitle = ({ type }: Props) => {
-  const titleText = (type: dataType) => {
-    switch (type) {
-      case "post":
-        return "내가 쓴 글"
-      case "comment":
-        return "내가 쓴 댓글"
-      case "bookmark":
-        return "북마크"
-      case "like":
-        return "좋아요 누른 글"
-    }
-  }
-  return (
-    <div className="w-full border-b border-b-black pb-[24px]">
-      <p>{titleText(type)}</p>
-    </div>
-  )
-}
-
 export const SearchBar = () => {
   return (
     <div className="flex gap-[16px]">
