+++ src/components/index.ts
@@ -1,11 +1,8 @@
-export * from "./button";
 export * from "./colorExtraction";
 export * from "./comment";
-export * from "./date";
+export * from "./common";
 // export * from './dialog';
-export * from "./input";
 export * from "./layout";
-export * from './modals';
 export * from './mypage';
 export * from './overlay';
 export * from "./postBookmark";
