--- src/components/input/Input.tsx
@@ -1,1 +1,0 @@
-export const Input = () => {};
