+++ src/components/mypage/PreviewEmpty.tsx
@@ -1,0 +1,7 @@
+export const PreviewEmpty = () => {
+  return (
+    <div className="w-full h-full flex contents-center px-[24px]">
+      <p className="text-black body-3">현재 데이터가 없습니다.</p>
+    </div>
+  )
+}
