+++ src/components/mypage/common/EmptyData.tsx
