+++ src/components/mypage/common/MypageTitle.tsx
@@ -1,0 +1,7 @@
+export const MypageTitle = () => {
+  return (
+    <div className="w-full  text-center pb-[24px]">
+      <h3 className="text-[32px] font-normal leading-[130%]">마이페이지</h3>
+    </div>
+  )
+}
