+++ src/components/mypage/common/index.ts
@@ -1,0 +1,3 @@
+export * from "./EmptyData"
+export * from "./MypageTitle"
+export * from "./MypageSubTitle"
