+++ src/components/mypage/index.ts
@@ -1,3 +1,5 @@
-export * from "./EmptyData"
-export * from "./MypageCommon"
-export * from "./PreviewDetail"
+export * from "./common"
+export * from "./PreviewBookmark"
+export * from "./PreviewComment"
+export * from "./PreviewLike"
+export * from "./PreviewPost"
