+++ src/components/mypage/preview.style.ts
@@ -1,0 +1,4 @@
+export const outerBoxStyle = "flex-column h-[130px]"
+export const innerBoxStyle = "w-full h-[64px] flex justify-between items-center px-[24px]"
+export const linkStyle = "text-black cursor-pointer body-3 hover:text-gray03"
+export const dateStyle = "text-gray02 body-3"
